Fix typo in Disney specular refl lobe creation
I figured this out by manually solo-ing lobes in the Material.create_disney function and comparing the render with the corresponding Rust render until I saw that the Disney specular lobe was coming out green colored. This isn't supposed to happen since specuar_tint=0 which should have made the specular white. Then I traced it down to this point, where the code didn't match the description in the comment, which explicitly states that "the color will be computed by the Disney fresnel function, so we just set it to white".